moved bkg stager to module
